requirements.txt
mysql-connector-python
pandas
python-decouple
requests
jupyterlab
